Added a setup.py
Setuptools is the standard way to package python projects.
Here's a basic file, it can be called directly to create a source or
binary package.
python setup.py sdist
python setup.py bdist
You can even create an rpm should you like with "python setup.py bdist_rpm".

MIT is the standard license for our open source stuff for now.

Markdown is not supported by PyPI so the readme needs to be in
reStructuredText to support both them and github, but I'll leave it as
it is for now.